# quiz-api/openapi.yaml
...
openapi: 3.0.3
info:
  title: Quiz API
  description: A RESTful API for managing quizzes and user progress
  version: 1.0.0
servers:
  - url: http://localhost:5000
    description: Development server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        username:
          type: string
          minLength: 3
          maxLength: 80
        email:
          type: string
          format: email
        full_name:
          type: string
          minLength: 1
        role:
          type: string
          enum: [admin, user]
        dob:
          type: string
          format: date
          description: Date of birth in DD/MM/YYYY format
        joined_at:
          type: string
          format: date-time
          readOnly: true
...
paths:
  /auth/register:
    post:
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password, email, full_name]
              properties:
                username:
                  type: string
                  minLength: 3
                password:
                  type: string
                  minLength: 6
                email:
                  type: string
                  format: email
                full_name:
                  type: string
                dob:
                  type: string
                  description: DD/MM/YYYY format
                role:
                  type: string
                  enum: [user]
      responses:
        "201":
          description: User registered successfully
        "400":
          description: Validation error or user already exists
  /auth/login:
    post:
      summary: Login to get access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                  user:
                    $ref: "#/components/schemas/User"
        "401":
          description: Invalid credentials
  /auth/me:
    get:
      summary: Get current user details
      security:
        - BearerAuth: []
      responses:
        "200":
          description: User details retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
...